requirements.txt
crewai>=0.108.0
crewai-tools>=0.37.0
anthropic>=0.42.0
pyodbc>=5.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-a2a[anthropic]>=0.5.0
flask>=2.0.0
